Move, add, remove, select nodes. Move expression. Cannot move/parenthesize selection

+++ core/src/train/chu/chu/model/Positioned.java
@@ -4,8 +4,8 @@ package train.chu.chu.model;
  * Created by Da-Jin on 6/5/2016.
  */
 public interface Positioned {
-    int getX();
-    void setX(int x);
-    int getY();
-    void setY(int x);
+    float getX();
+    void setX(float x);
+    float getY();
+    void setY(float x);
 }
